+++ backend/src/main/java/org/example/livecoinwatch/response/CoinsSingle.kt
@@ -1,7 +1,9 @@
 package org.example.livecoinwatch.response
 
+import java.math.BigDecimal
+
 data class CoinsSingle(
-val rate: Double,
+val rate: BigDecimal,
 val volume: Long,
 val cap: Long,
 val liquidity: Long,
